% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Calculate the sample mean}
\usage{
sample_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The sample mean of \code{x}.
}
\description{
Calculates the arithmetic mean of a numeric vector.
}
\examples{
sample_mean(c(1, 2, 3, 4, 5))

}
