% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_variance.R
\name{sample_variance}
\alias{sample_variance}
\title{Calculate the sample variance}
\usage{
sample_variance(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The sample variance of \code{x}.
}
\description{
Calculates the sample variance of a numeric vector.
}
\examples{
sample_variance(c(1, 2, 3, 4, 5))

}
