% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The standard error of \code{x}.
}
\description{
Calculates the standard error of a numeric vector.
}
\examples{
standard_error(c(1, 2, 3, 4, 5))

}
